Added Part4 user Authentication and testing

+++ part_4/controllers/blogController.js
@@ -1,15 +1,18 @@
 // API routes to MongoDB
 const blogRouter = require('express').Router()
 const blogObj    = require('../models/blogSchema')
-const logger     = require('../utils/logger')
+const userObj    = require('../models/userSchema')
+const jwt        = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
     const result = await blogObj.find({})
     return response.json(result)
 })
 
+
 blogRouter.post('/', async (request, response) => {
     const body = request.body
+    const user = request.user
 
     if(!body.likes){
         body.likes = 0
@@ -23,42 +26,55 @@ blogRouter.post('/', async (request, response) => {
         })
     }
 
+    
     const newBlog = new blogObj({
         title:  body.title,
         author: body.author,
         url:    body.url,
-        likes:  body.likes
+        likes:  body.likes,
+        user:   user._id
     })
 
-    const result = await newBlog.save()
+    const savedBlog = await newBlog.save()
 
-     return response.json(result)
+    user.blog = user.blog.concat(savedBlog._id)
+    await user.save()
+    return response.json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    const result = await blogObj.findByIdAndDelete(request.params.id)
+    const user = request.user
 
-    if(result){
-        response.status(204).send(`${request.params.id} deleted`)
+    const blogToDelete = await blogObj.findById(request.params.id)
+    if (!blogToDelete){
+        return response.status(400).json({message: `${request.params.id} not found. No action taken.` })
+    } else if(blogToDelete.user.toString() === user.id.toString()){
+        await blogObj.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
     } else {
-        response.send(`${request.params.id} not found. No action taken`)
+        return response.status(401).json({error: `${request.params.id} blog doesnt belong to user.`})
     }
+
 })
 
 blogRouter.put('/:id', async (request, response) => {
     const body = request.body
+    const user = request.user
+
     const newBlog = {
         author: body.author,
         likes: body.likes,
         title: body.title,
         url: body.url
     }
-    const result = await blogObj.findOneAndUpdate(request.params.id, newBlog, {new: true} )
-
-    if(result){
-        response.status(204).json(result)
+    const blogToUpdate = await blogObj.findById(request.params.id)
+    if(!blogToUpdate){
+        return response.status(400).json({ message: `${request.params.id} not found. No action taken`})
+    } else if (blogToUpdate.user.toString() === user.id.toString()){
+        await blogObj.findOneAndUpdate(request.params.id, newBlog, {new: true} )
+        return response.status(204).end()
     } else {
-        response.send(`${request.params.id} not found. No action taken`)
+        return response.status(401).json({ error: `${request.params.id} blog doesnt belong to user`})
     }
 })
